- clean up db only if it has more than 30 records - schedule next in-app-message based on the interval set - check for start and end time before showing in-app-message - queue behaviour added to avoid waiting time for next in-app-message

+++ android-sdk/src/main/java/com/blueshift/inappmessage/InAppMessageStore.java
@@ -152,10 +152,12 @@ public class InAppMessageStore extends BlueshiftBaseSQLiteOpenHelper<InAppMessag
     }
 
     public void clean() {
-        String timestamp = String.valueOf(System.currentTimeMillis() / 1000);
-        String whereClause = FIELD_EXPIRES_AT + "<?";
-        String[] selectionArgs = new String[]{timestamp};
+        if (getTotalRecordCount() > 30) {
+            String timestamp = String.valueOf(System.currentTimeMillis() / 1000);
+            String whereClause = FIELD_EXPIRES_AT + "<?";
+            String[] selectionArgs = new String[]{timestamp};
 
-        deleteAll(whereClause, selectionArgs);
+            deleteAll(whereClause, selectionArgs);
+        }
     }
 }
